refactor: increase bot separation into files
