+++ telegram-bot/bot/core.py
@@ -1,4 +1,6 @@
-from aiogram import Bot
+from aiogram import Bot, Dispatcher
 from decouple import config
 
 bot = Bot(token=config("BOT_TOKEN"))
+dp = Dispatcher()
+
