+++ telegram-bot/bot/handlers.py
@@ -6,23 +6,19 @@
 #       start.py
 #       booking.py
 #       ...
-#   instance.py
+#   core.py
 
-from aiogram import F, Router
-from aiogram.types import Message
-from client.http_client import FortunaAPIClient
+from aiogram import Router
+from client.http_client import APIClient
 from decouple import config
 
-from bot.instance import bot
-
 message_router = Router()
-api_client = FortunaAPIClient(config("FORTUNA_API_URL"))
+api_client = APIClient(config("FORTUNA_API_URL"))
 
-
-@message_router.message(F.text)
-async def handle_message(message: Message) -> None:
-    await api_client.save_message(
-        user_id=message.from_user.id,
-        text=message.text
-    )
-    await bot.send_message(message.chat.id, "Сообщение сохранено!")
+# @message_router.message(F.text)
+# async def handle_message(message: Message) -> None:
+#     await api_client.save_message(
+#         user_id=message.from_user.id,
+#         text=message.text
+#     )
+#     await bot.send_message(message.chat.id, "Сообщение сохранено!")
